requirements.txt
requests
aiogram
pyaudio
pydub
numpy
setuptools
scipy
python-dotenv
speechrecognition